refactored modifiers, stas and skills to use decimal values; defined Size stat and implemented Carrying Capacity

+++ Xethya/Common/Interfaces/IWithModifiers.cs
@@ -20,6 +20,6 @@ namespace Xethya.Common.Interfaces
         /// <summary>
         /// Returns the sum of each modifier's value.
         /// </summary>
-        int ModifierSum { get; }
+        decimal ModifierSum { get; }
     }
 }
